organizing and beginning to build out search
