+++ src/Web/src/App.js
@@ -7,7 +7,7 @@ import {
 
 import './App.css';
 import Navigation from './components/Navigation';
-import Directory from './components/Directory';
+import Directory from './components/DirectoryComponents/Directory';
 import Profile from './components/Profile';
 
 function App() {
@@ -15,13 +15,13 @@ function App() {
     <div className="App">
       <Navigation />
       <Router>
-        {/* <Switch> */}
+          <Switch>
           <div className="body">
             <Route path="/:parameters" render={() => <Directory />} />
             {/* add :id when we have api feeding it */}
             <Route exact path="/profile" component={Profile} />
           </div>
-        {/* </Switch> */}
+          </Switch>
       </Router>
     </div>
   );
